Create message form live component
